# feat: cache common `StringDoc` types