--- a/Src/CSharpier/DocTypes/Doc.cs
+++ b/Src/CSharpier/DocTypes/Doc.cs
@@ -12,10 +12,7 @@ internal abstract class Doc
         return DocSerializer.Serialize(this);
     }
 
-    public static implicit operator Doc(string value)
-    {
-        return new StringDoc(value);
-    }
+    public static implicit operator Doc(string value) => StringDoc.Create(value);
 
     public static NullDoc Null => NullDoc.Instance;