--- a/Src/CSharpier/SyntaxPrinter/Token.cs
+++ b/Src/CSharpier/SyntaxPrinter/Token.cs
@@ -116,7 +116,7 @@ internal static class Token
         }
         else
         {
-            docs.Append(syntaxToken.Text);
+            docs.Append(StringDoc.Create(syntaxToken));
         }
 
         if (!skipTrailingTrivia)