… Src/CSharpier/DocTypes/StringDoc.cs