@belav 2025-02-22
When I saw the title of this PR I was picturing the code making use of something like `Strings.Comma` but looking at the code again I think almost all of these come from `Poken.PrintSyntaxToken`. So I think you'd still end up with a switch statement there. Not sure if it is worth changing the code to make use of `StringDoc.SpaceString`.

> Is it worth exploring caching all new StringDoc types? There are a total of 674 unique StringDoc symbols in the example benchmark, I wonder how many unique symbols are in a real world 1M loc project. Perhaps a cache for each file or using a static ConditionalWeakTable.
I can see this causing a memory leak if implemented incorrectly😅

I would be a bit worried it would cause problems but if you format everything within `csharpier-repos` I imagine a memory leak would be pretty obvious.

Maybe it is better to include the common symbols like you mentioned and keep it as a simple switch.

> A future change could look into creating a SpanDoc type that contains SyntaxToken.Span, this is used to copy the relevant section of the original code into the new text without allocating.

I did some poking around and it looks like `SyntaxToken.Span` is for the start and end position of that Token in the class. `SyntaxToken.Text`  is a bit confusing to me. Poking around with rider it looked like it would end up calling `GreenNode.ToString()` on every call to `SyntaxToken.Text` which writes to a `StringBuilder`.

When debugging though, it ends up in an interal class `SyntaxToken` which eventually calls `SyntaxFacts.GetText(SyntaxKind)`. That is a giant switch statement which returns a string, and would probably be a starting place for caching more things. It could also operate off of `SyntaxKind` instead of `string`

https://github.com/dotnet/roslyn/blob/main/src/Compilers/CSharp/Portable/Syntax/SyntaxKindFacts.cs#L1363

I also noticed that `GreenNode` makes use of `PooledStringBuilder`. It is internal, but provides an alternative to passing around and/or creating instances of `StringBuilder`. I'm sure it could be adopted for Stack as well. https://github.com/dotnet/roslyn/blob/main/src/Dependencies/PooledObjects/PooledStringBuilder.cs