@TimothyMakkison 2025-02-23
So it looks like roslyn is smart, and when it parses a new symbol or token ie an identifier, it will use an internal cache called a `StringTable` and check if a string matching the text exists. If it exists it will just return a reference to this existing string instead of allocating a duplicate.

For example calling `SyntaxToken.Text` on an indentifier calls `Node.ToString` which calls `SyntaxIdentifier.Text` which returns its internal textfield. At no point does it allocate a new string.

This means that creating a `SpanDoc` is completely unneeded. Like you, I had also assumed that `GreenNode` was allocating strings like crazy 😄 it shows how optimised roslyn is.

>  eventually calls SyntaxFacts.GetText(SyntaxKind). That is a giant switch statement which returns a string, and would probably be a starting place for caching more things. It could also operate off of SyntaxKind instead of string

This is definitely worth exploring. Should I close this PR in favour of one caching all `SyntaxKind` text?
I'll see if I can rerun my above experiment and see how many of the `StringDoc` values are C# symbols vs user defined words.

> also noticed that GreenNode makes use of PooledStringBuilder. It is internal, but provides an alternative to passing around and/or creating instances of StringBuilder. I'm sure it could be adopted for Stack as well.

Yeah microsoft loves using these 😅 I did wonder if `DocPrinter.Output` and `PropagateBreaks.RunOn.alreadyVisitedSet`would be candidates for object pooling. I need to check the guidance on when to use pooling.
I'm going to make a pr that uses `ValueStringBuilder` for `SeparatedSyntaxList`, I tried doing this before in #1490 but opted for #1491.

>  I'm sure it could be adopted for Stack as well.

I don't think this is currently worth doing, after #1491 csharpier doesn't construct many `Stack<T>` collections. The only improvement would be to use `ValueListBuilder` as a stack as it has instant indexing.