@TimothyMakkison 2025-02-24
Oh yeah, csharpier should definitely be using `SyntaxToken.Kind` and `SyntaxNode.Kind` instead of `is`. I feel like `Node.Print` would see a massive speed up from it. It might be as easy as using the filename of each printer. I am a little worried it won't handle inheritance of nodes like `is`, although I assume the roslyn team have considered this, as they use this pattern internally a lot.