@TimothyMakkison 2025-03-15
- Added `ConditionalWeakTable`
- Renamed `ToStringDoc` to `Create`

> Reading up on ConditionalWeakTable I... think it would be okay?

> When there are no more references to the key outside of the table, then the entry can be removed from the table. The StringDoc that is the value in the table is returned from these methods, so it will be referenced in other parts of the code. The key isn't really referenced in the code beyond the call to ToStringDoc so it will be eligible to be removed from the table before we are done with the StringDoc.

> So assuming that removing an entry from the table doesn't GC the values in the table that may still be referenced then I think this won't cause any issues.


I agree, this should be perfect for our needs here.


> Assuming the 15% is even close to accurate that is a pretty significant boost and I think worth it.

I suspect it was just a series of freak results, who knows perhaps its accurate 🤷 New changes saves 3.2 MB on `Test` benchmark.

I've since deleted the original benchmarks but it looks like `WeakTable` saves 0.6 MB and `StringDoc` saves 2.6 MB. I might have misread the original numbers or the `SyntaxToken` helped more than I thought 🤔


#### Benchmarks (Done on battery power, timing is 100% random)
### Main

| Method                        | Mean     | Error   | StdDev   | Median   | Gen0      | Gen1      | Allocated |
|------------------------------ |---------:|--------:|---------:|---------:|----------:|----------:|----------:|
| Default_CodeFormatter_Tests   | 137.0 ms | 2.71 ms |  6.95 ms | 136.7 ms | 3000.0000 | 1000.0000 |  32.02 MB |
| Default_CodeFormatter_Complex | 299.9 ms | 7.65 ms | 22.56 ms | 307.2 ms | 5000.0000 | 2000.0000 |  56.41 MB |


### String Doc

| Method                        | Mean     | Error   | StdDev  | Gen0      | Gen1      | Allocated |
|------------------------------ |---------:|--------:|--------:|----------:|----------:|----------:|
| Default_CodeFormatter_Tests   | 135.4 ms | 2.66 ms | 4.58 ms | 2000.0000 | 1000.0000 |  29.37 MB |
| Default_CodeFormatter_Complex | 276.1 ms | 5.35 ms | 7.49 ms | 5000.0000 | 2000.0000 |  52.79 MB |


### StringDoc WeakTable
| Method                        | Mean     | Error   | StdDev   | Gen0      | Gen1      | Allocated |
|------------------------------ |---------:|--------:|---------:|----------:|----------:|----------:|
| Default_CodeFormatter_Tests   | 175.1 ms | 3.49 ms |  8.69 ms | 2000.0000 | 1000.0000 |  28.77 MB |
| Default_CodeFormatter_Complex | 279.3 ms | 7.84 ms | 22.74 ms | 4000.0000 | 2000.0000 |  51.27 MB |