@TimothyMakkison 2025-03-16
### Benchmarks (more accurate this time)
#### Before
| Method                        | Mean     | Error   | StdDev  | Gen0      | Gen1      | Allocated |
|------------------------------ |---------:|--------:|--------:|----------:|----------:|----------:|
| Default_CodeFormatter_Tests   | 102.1 ms | 2.20 ms | 6.45 ms | 2000.0000 | 1000.0000 |  31.07 MB |
| Default_CodeFormatter_Complex | 189.8 ms | 2.93 ms | 5.43 ms | 5000.0000 | 2000.0000 |  54.57 MB |

#### Before 10x
| Method                        | Mean    | Error    | StdDev   | Gen0       | Gen1       | Gen2      | Allocated |
|------------------------------ |--------:|---------:|---------:|-----------:|-----------:|----------:|----------:|
| Default_CodeFormatter_Tests   | 1.192 s | 0.0237 s | 0.0325 s | 37000.0000 | 23000.0000 | 7000.0000 |  317.5 MB |
| Default_CodeFormatter_Complex | 2.405 s | 0.0464 s | 0.1310 s | 62000.0000 | 35000.0000 | 9000.0000 | 545.37 MB |


#### After
| Method                        | Mean      | Error    | StdDev    | Gen0      | Gen1      | Allocated |
|------------------------------ |----------:|---------:|----------:|----------:|----------:|----------:|
| Default_CodeFormatter_Tests   |  98.11 ms | 1.951 ms |  4.785 ms | 2000.0000 | 1000.0000 |  28.59 MB |
| Default_CodeFormatter_Complex | 180.12 ms | 5.030 ms | 14.832 ms | 4000.0000 | 2000.0000 |  49.49 MB |

#### After 10x

| Method                        | Mean    | Error    | StdDev   | Gen0       | Gen1       | Gen2      | Allocated |
|------------------------------ |--------:|---------:|---------:|-----------:|-----------:|----------:|----------:|
| Default_CodeFormatter_Tests   | 1.052 s | 0.0209 s | 0.0382 s | 33000.0000 | 19000.0000 | 6000.0000 | 287.69 MB |
| Default_CodeFormatter_Complex | 2.182 s | 0.0435 s | 0.0817 s | 56000.0000 | 33000.0000 | 9000.0000 | 495.07 MB |